Merge Calendar.service into Calendar.controller
At the origin I made Calendar.service to extract all the logic from
Calendar.controller. In doing so, the only responsibility of the
controller was to manipulate and interact with the template.

However, there were no benefits to use a service here, for 2 reasons:
- the singleton pattern was not needed
- the service had no dependencies (e.g. $http)

My first idea was to transform the service to a simple object in a
separate file and import it in the controller. In doing so, you don’t
have to register the service and use the Dependency Injection system.

Finally I decided to merge the whole service into the controller
because they were closely related and there were no other controllers
depending on it.